Quiz Entity und QuestionEntity erstellt und Mapper für die konvertierung vom Domain Objekt zum Datenbank Objekt erstellt + getestet + Builder im Test

+++ src/main/java/de/propra/exam/DTO/QuestionDTO.java
@@ -4,7 +4,7 @@ public class QuestionDTO {
     private String title;
     private String task;
     private String solution;
-    private Integer points;
+    private Double points;
     private String type;
 
     public String getType() {
@@ -39,11 +39,11 @@ public class QuestionDTO {
         this.solution = solution;
     }
 
-    public Integer getPoints() {
+    public Double getPoints() {
         return points;
     }
 
-    public void setPoints(Integer points) {
+    public void setPoints(Double points) {
         this.points = points;
     }
 }
